requirements.txt
dropbox==10.2.0
Flask==1.1.2
Flask-HTTPAuth==4.1.0
openpyxl==3.0.4
Pillow==7.2.0
PyJWT==1.7.1
pytest==5.4.3
pytest-cov==2.10.0
-e .
